robot/src/jetbot_mini/scripts/requirements.txt
fastapi==0.111.0
keras==3.3.3
librosa==0.10.2
matplotlib==3.8.4
mediapipe==0.10.14
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
pydantic==2.7.3
scikeras==0.13.0
scikit_learn==1.4.2
tensorflow==2.16.1
uvicorn==0.30.1
